requirements.txt
click==8.1.7
dnspython==2.6.1
markdown-it-py==3.0.0
mdurl==0.1.2
NetfilterQueue @ git+https://github.com/kti/python-netfilterqueue@542159a6312726ed49c7cf3639c578d96fc0d1fd
prompt-toolkit==1.0.14
Pygments==2.18.0
PyInquirer==1.0.3
python-dotenv==1.0.1
regex==2024.9.11
rich==13.8.1
scapy==2.5.0
shellingham==1.5.4
six==1.16.0
typer==0.12.5
typing_extensions==4.12.2
wcwidth==0.2.13
